% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_gdaphen_analysis.R
\name{any_categorical_variable_discrimination_byPhenoParams.function}
\alias{any_categorical_variable_discrimination_byPhenoParams.function}
\title{MFA extraction - Creation of plots input dataframes}
\usage{
any_categorical_variable_discrimination_byPhenoParams.function(
  InputdataPCAMix,
  nameExperiment,
  catVarName,
  phenoColourDf,
  indSexCond,
  InputDf,
  qualiVars,
  nameModel
)
}
\arguments{
\item{InputdataPCAMix}{The input is the output from the mfa.function, the object:
\link{res_PCAmixData_}nameOutput}

\item{nameExperiment}{The name of the experiment together with the number of variables we included.
if is all model, allthe variables. Instead if is the selected model is less variables.
ex:
\enumerate{
\item nameExperiment <- paste0(name,"_", length(colnames(phenoSocial_4analysis)), " variables"); #all variables
\item nameExperiment <- paste0(name,"_", length(colnames(phenoAll_4analysis_woHighCor_sel1)), " variables"); · without highly correlated
\item nameExperiment <- paste0(name,"_", " more informative variables ",length(colnames(parameters_more_30perc_informative_Genotype)) ) #selected variables
}}

\item{catVarName}{Column name of the qualitative
variable or categorical variable that is your dependant
variable, the one for which one you want to identify the most discriminative variables}

\item{phenoColourDf}{Dataframe containing at least 2 columns,
1: "labels", containing the possible categories. Ex Genotype het, wt, homo
2: "colPalette" hex colour associated to each varible.
Ex: for label="heterozygous", use colPalette="#8B0000"}

\item{indSexCond}{Samples data color asignation for each Individual point depending
on genotypes, sex, treatments is created by the user.
an example of how to create it is in the auxiliar function
' indSexCond.colour_definition.function.}

\item{InputDf}{Is always the phenoSocial_4analysis file created in the preproccesing steps
condition_definition.function.}

\item{qualiVars}{Number of qualitative variables used on the model.}

\item{nameModel}{Shrt naming for the model you are using.
Ex: "fullMdl" for the model including all variables;"Sel1" for the model not including the
highly correlated variables, "Sel30" for the model including only the variables contributing
more thana 30\%}
}
\value{
6 dataframe and several variables that contain the information for plotting
namesVariablesDf, df1, df2, df3,df4, df5, df6, MFA1, MFA2, MFA3,nbVarF, nbVarTotalF'
}
\description{
MFA extraction - Creation of plots input dataframes
}
\examples{
any_categorical_variable_discrimination_byPhenoParams.function(InputdataPCAMix,nameExperiment,catVarName,phenoColourDf,indSexCond,phenoSocial_4analysis,qualiVars,nameModel)

}
