% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_gdaphen_analysis.R
\name{any_categorical_variable_discrimination_selection_most_informative_params_more_30.function}
\alias{any_categorical_variable_discrimination_selection_most_informative_params_more_30.function}
\title{More informativeVars in PCA - Identification of Sel30 variables}
\usage{
any_categorical_variable_discrimination_selection_most_informative_params_more_30.function(
  InputdataPCAMix,
  catVarName,
  phenoColourDf
)
}
\arguments{
\item{InputdataPCAMix}{The input is the output from the mfa.function, the object: \link{res_PCAmixData_}nameOutput}

\item{catVarName}{Column name of the qualitative
variable or categorical variable that is your dependant
variable, the one for which one you want to identify the most discriminative variables}

\item{phenoColourDf}{Dataframe containing at least 2 columns,
1: "labels", containing the possible categories. Ex Genotype het, wt, homo
2: "colPalette" hex colour associated to each varible.
Ex: for label="heterozygous", use colPalette="#8B0000"}
}
\value{
A dataframe, like dataInputPCA but containing only the columns
or variables thata re contributing more than a 30\% to the discrimination
Note: Sex and genotype are always preselected.These columns do not follow
the order the experimentator provided in phenoSocial_4analysis. The output has a prefix
catVarName, and followed by "_sel_Vars_contribution_more_30perc_plus_categorical"
}
\description{
More informativeVars in PCA - Identification of Sel30 variables
}
\examples{
# for discrimination of  Genotype:
any_categorical_variable_discrimination_selection_most_informative_params_more_30.function(res_PCAmixData_phenoAll_4analysis_woHighCor_sel1,"Genotype",phenoColourDf)
# for discrimination of  Sex:
any_categorical_variable_discrimination_selection_most_informative_params_more_30.function(res_PCAmixData_phenoAll_4analysis_woHighCor_sel1,"Sex",phenoColourDf)

}
