% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_preProcessing_functions.R
\name{condition_definition.function}
\alias{condition_definition.function}
\title{Preformatting- Annonymization adn correlation}
\usage{
condition_definition.function(
  inputDf,
  conditionNameType,
  catVariables_initial_colNb
)
}
\arguments{
\item{conditionNameType}{"perRowNumber"  when we want to annonymize or run Gdaphen in more than
1 model or one model with different ages/treatments ,when the name labels are long, etc.
Instead we will use conditionNameType=Ind when is only one model, with one age:treatment data
By default recommended conditionNameType=perRowNumber}

\item{catVariables_initial_colNb}{Number of qualitative variables on InputDf, perform a
head(InputDf) to get this number.}

\item{InputDf}{Is  the phenoInput file , orginal file after imputation if is needed,
of the data provided by the experimentator.}
}
\value{
Several objects, the input for the full model phenoSocial_4analysis,
the imput for the selected model without highly corr variables, noted as Sel1,
phenoAll_4analysis_woHighCor_sel1. Both with and without Genotype.
The correlationMatrix, and other parameters for downstream functions.
pheno: Is the InputDf, assigned to anew name to avoid touching the input data.
}
\description{
Preformatting- Annonymization adn correlation
}
\examples{
condition_definition.function(phenoInput,"perRowNumber");
condition_definition.function(phenoInput,"Ind"); 
condition_definition.function(phenoInput,"Ind",5);
}
