% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_gdaphen_visualization.R
\name{cumVariance_Plot.function}
\alias{cumVariance_Plot.function}
\title{plot0 - Cummulative variance plot over 10 components PCA}
\usage{
cumVariance_Plot.function(ResPCAmixData, nameOutput)
}
\arguments{
\item{ResPCAmixData}{Input file is the output from MFA.function.}

\item{nameOutput}{Short naming for the output of the plot.}
}
\value{
The plot 0 is generated in the cumulativeVariance folder
}
\description{
plot0 - Cummulative variance plot over 10 components PCA
}
\examples{
cumVariance_Plot.function(res_PCAmixData_phenoAll_4analysis,"res_PCAmixData_phenoAll_4analysis");
cumVariance_Plot.function(res_PCAmixData_phenoAll_4analysis_woHighCor_sel1,"res_PCAmixData_phenoAll_4analysis_woHighCor_sel1");
cumVariance_Plot.function(res_PCAmixData_parameters_more_30perc_informative_Genotype,"res_PCAmixData_parameters_more_30perc_informative_Genotype");
cumVariance_Plot.function(res_PCAmixData_parameters_more_30perc_informative_woGenotype,"res_PCAmixData_parameters_more_30perc_informative_woGenotype");

}
