% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_preProcessing_functions.R
\name{indSexCond.colour_definition.function}
\alias{indSexCond.colour_definition.function}
\title{Preformatting colour per sample.}
\usage{
indSexCond.colour_definition.function(
  pheno,
  catVar_selected,
  name_catVar_selected
)
}
\arguments{
\item{pheno}{Is  the InputDf file , orginal file after imputation if is needed,
of the data provided by the experimentator.}

\item{catVar_selected}{name of the categorical variables that are selected for the analyses}

\item{name_catVar_selected}{Rename if you want the columns names representing the categorical variable selected.}
}
\value{
Dataframe with the name fo the samples, and the color asignation for each Individual point depending
on the categorical variables levels. Ex: depending on genotypes, sex,
treatments. This function is advised to be modified by the user as you may have adifferent experimental
}
\description{
Preformatting colour per sample.
}
\examples{
indSexCond.colour_definition.function(pheno, catVar_selected,catVar_selected);

}
