% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_gdaphen_analysis.R
\name{linear_equation_phenoVars_discrimination.function}
\alias{linear_equation_phenoVars_discrimination.function}
\title{Linear equation}
\usage{
linear_equation_phenoVars_discrimination.function(
  InputdataPCAMix,
  nameExperiment
)
}
\arguments{
\item{InputdataPCAMix}{The input is the output from the mfa.function, the object:
\link{res_PCAmixData_}nameOutput}

\item{nameExperiment}{The name of the experiment together with the number of variables we included.
if is all model, allthe variables. Instead if is the selected model is less variables.
ex:
\enumerate{
\item nameExperiment <- paste0(name,"_", length(colnames(phenoSocial_4analysis)), " variables"); #all variables
\item nameExperiment <- paste0(name,"_", length(colnames(phenoAll_4analysis_woHighCor_sel1)), " variables"); · without highly correlated
\item nameExperiment <- paste0(name,"_", " more informative variables ",length(colnames(parameters_more_30perc_informative_Genotype)) ) #selected variables
}}
}
\value{
A dataframe saved in excel with the linear equation computed.
}
\description{
Linear equation
}
\examples{
linear_equation_phenoVars_discrimination.function(InputdataPCAMix,nameExperiment);

}
