% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_gdaphen_analysis.R
\name{mfa.function}
\alias{mfa.function}
\title{mfa.function - Function to perform Multiple Factor Analysis for mixed data (MFAmix)}
\usage{
mfa.function(dataInputPCA, catVarNb, nameOutput)
}
\arguments{
\item{dataInputPCA}{Datframe where each row represent an animal/pair of animal measurements, and the columns represents
qualitative/quantitative parameters measurements. Data should be complete no NAs accepted,
if needed imputation strategies should be used before running this function.}

\item{catVarNb}{Number of categorical variables contained in dataInputPCA. Ex use 2 if Genotype and Sex  are present.}

\item{nameOutput}{Sufix Name specific from this project to add to the name of the output of the function}
}
\value{
A list of dataframes object: \link{res_PCAmixData_}nameOutput
}
\description{
mfa.function - Function to perform Multiple Factor Analysis for mixed data (MFAmix)
}
\examples{
# for discrimination of  Genotype:
mfa.function(parameters_more_30perc_informative_Genotype,2, "parameters_more_30perc_informative_Genotype");
# after removing Genotype, which variables are the most discriminative?
mfa.function(parameters_more_30perc_informative_woGenotype_Genotype,1, "parameters_more_30perc_informative_woGenotype");
# for discrimination of  Sex:
mfa.function(parameters_more_30perc_informative_Sex,2, "parameters_more_30perc_informative_Sex");

}
