% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_gdaphen_analysis.R
\name{parameters_more_30perc_informative_creation_InputDf.function}
\alias{parameters_more_30perc_informative_creation_InputDf.function}
\title{Parameters_more_30perc_informative - Creating the Sel30 model}
\usage{
parameters_more_30perc_informative_creation_InputDf.function(
  phenoSocial_4analysis,
  sel_Vars_contribution_more_30perc_plus_categorical,
  nameOutput
)
}
\arguments{
\item{phenoSocial_4analysis}{The input is the output from the mfa.function, the object: \link{res_PCAmixData_}nameOutput}

\item{sel_Vars_contribution_more_30perc_plus_categorical}{Output of the function
informative_params_more_30.function
variable or categorical variable that is your dependant
variable, the one for which one you want to identify the most discriminative variables}

\item{nameOutput}{Sufix Name specific from this project to add to the name of the output of the function}
}
\value{
A dataframe, like phenoSocial_4analysis but containing only the columns
or variables that are contributing more than a 30\% to the discrimination
Contrary to the dataframe produced by
informative_params_more_30.function
in this case, the columns are following the order specified by the experimentator in phenoSocial_4analysis
These columns do not follow the order the experimentator provided in the dataInputPCA.
The output has  aprefix= "parameters_more_30perc_informative_" and as  a sufix nameOutput
}
\description{
Parameters_more_30perc_informative - Creating the Sel30 model
}
\examples{
# for discrimination of  Genotype:
parameters_more_30perc_informative_creation_InputDf.function(phenoSocial_4analysis,Genotype_sel_Vars_contribution_more_30perc_plus_categorical,"Genotype");
# for discrimination of  Sex:
parameters_more_30perc_informative_creation_InputDf.function(phenoSocial_4analysis,Sex_sel_Vars_contribution_more_30perc_plus_categorical,"Sex");

}
