% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_gdaphen_visualization.R
\name{plot6_qualitative_vars_discrimination_PCA.function}
\alias{plot6_qualitative_vars_discrimination_PCA.function}
\title{plot6 -	Quantitative variables distribution in PCA}
\usage{
plot6_qualitative_vars_discrimination_PCA.function(
  name,
  catVarName,
  nameExperiment,
  nrowLegend,
  Cat_variableToShow,
  nameModel
)
}
\arguments{
\item{name}{name for the plot or specific to the session}

\item{catVarName}{Column name of the qualitative
variable or categorical variable that is your dependant
variable, the one for which one you want to identify the most discriminative variables}

\item{nameExperiment}{The name of the experiment together with the number of variables we included.
if is all model, allthe variables. Instead if is the selected model is less variables.
ex:
\enumerate{
\item nameExperiment <- paste0(name,"_", length(colnames(phenoSocial_4analysis)), " variables"); #all variables
\item nameExperiment <- paste0(name,"_", length(colnames(phenoAll_4analysis_woHighCor_sel1)), " variables"); · without highly correlated
\item nameExperiment <- paste0(name,"_", " more informative variables ",length(colnames(parameters_more_30perc_informative_Genotype)) ) #selected variables
}}

\item{nrowLegend}{Number of rows the legend should be split to, to format properly for publications.}

\item{Cat_variableToShow}{name of the specific categorical variable to show. Specify isntead  "All"
to show all in the plot.}

\item{nameModel}{Short naming for the model you are using.
Ex: "fullMdl" for the model including all variables;"Sel1" for the model not including the
highly correlated variables, "Sel30" for the model including only the variables contributing
more thana 30\%}
}
\value{
The plot 6 is generated in the plot6 folder in three different kind of organizations
to select the better for publications (landscape, vertical, 1 plot alone or all plots together.
}
\description{
plot6 -	Quantitative variables distribution in PCA
}
\examples{
# show the "Genotype" variable
plot6_qualitative_vars_discrimination_PCA.function(name,catVarName,nameExperiment,4,"Genotype",nameModel);
# show all the categorical variables . Ex. Sex and Genotype
plot6_qualitative_vars_discrimination_PCA.function(name,catVarName,nameExperiment,4,"All",nameModel);

}
