% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_preProcessing_functions.R
\name{random_colour_assignation.function}
\alias{random_colour_assignation.function}
\title{Preformatting- random colour assignation}
\usage{
random_colour_assignation.function(phenoColourDf)
}
\arguments{
\item{phenoColourDf}{Dataframe containing at least 2 columns,
1 "labels", containing the possible categories. Ex Genotype het, wt, homo
2 "colPalette" hex colour associated to each varible.
Ex for label="heterozygous", use colPalette="#8B0000"}
}
\value{
Several objects, the input for the full model phenoSocial_4analysis,
the imput for the selected model without highly corr variables, (noted as Sel1),
phenoAll_4analysis_woHighCor_sel1. Both with and without Genotype.
The correlationMatrix, and other parameters for downstream functions.
}
\description{
Preformatting- random colour assignation
}
\examples{
random_colour_assignation.function(phenoColourDf),

}
