% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_gdaphen_visualization.R
\name{scatterplot3d.function}
\alias{scatterplot3d.function}
\title{plot3 - Individuals observations in 3D-PCA}
\usage{
scatterplot3d.function(
  input_scatterplot3d,
  name,
  catVarName,
  nameExperiment,
  colorByColumn,
  colPalette_toPlot,
  height,
  width,
  m,
  ncolLegend,
  nameModel
)
}
\arguments{
\item{input_scatterplot3d}{input df dataframe o df3 or further customized}

\item{name}{name for the plot or specific to the session}

\item{catVarName}{Column name of the qualitative
variable or categorical variable that is your dependant
variable, the one for which one you want to identify the most discriminative variables}

\item{nameExperiment}{The name of the experiment together with the number of variables we included.
if is all model, allthe variables. Instead if is the selected model is less variables.
ex:
\enumerate{
\item nameExperiment <- paste0(name,"_", length(colnames(phenoSocial_4analysis)), " variables"); #all variables
\item nameExperiment <- paste0(name,"_", length(colnames(phenoAll_4analysis_woHighCor_sel1)), " variables"); · without highly correlated
\item nameExperiment <- paste0(name,"_", " more informative variables ",length(colnames(parameters_more_30perc_informative_Genotype)) ) #selected variables
}}

\item{colorByColumn}{If several conditions can be plotted. Ex Genotype, Genotype+ treatments, etc.
identify here the name of the variable to plot, is the name of the column from df3 that you want to plot.}

\item{colPalette_toPlot}{Column name from df3 that have the color assignation to plot the columns condition4Plot}

\item{height}{height ize in inches for the plot}

\item{width}{width size in inches for the plot}

\item{m}{layout in a matrix to organise the plots for plotting}

\item{ncolLegend}{Number of columns the legend should be split to, to format properly for publications.}

\item{nameModel}{Short naming for the model you are using.
Ex: "fullMdl" for the model including all variables;"Sel1" for the model not including the
highly correlated variables, "Sel30" for the model including only the variables contributing
more thana 30\%}
}
\value{
The 3D plots similar to the 2Dimension plot3 is generated in the plot3 folder in three different kind of organizations
to select the better for publications (landscape, vertical, 1 plot alone or all plots together.
}
\description{
plot3 - Individuals observations in 3D-PCA
}
\examples{
#run for all the genotypes comparison
m <- rbind(c(1,1,2))
scatterplot3d.function(input_scatterplot3d.all,name, catVarName, nameExperiment,"Genotype","colPalette_cond_Geno",5.1,8,m,1,nameModel)
scatterplot3d.function(input_scatterplot3d.all,name, catVarName, nameExperiment,"cond_Geno_Sex","colPalette_cond_Geno_Sex",5.1,8,m,1,nameModel)

}
