% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_gdaphen_analysis.R
\name{stats_cosineDistance_plot1.function}
\alias{stats_cosineDistance_plot1.function}
\title{Cosine Distance - Function to calculate cosine simmilarity distances}
\usage{
stats_cosineDistance_plot1.function(
  catVarName,
  similarityThreshold,
  nameExperiment
)
}
\arguments{
\item{catVarName}{Column name of the qualitative
variable or categorical variable that is your dependant
variable, the one for which one you want to identify the most discriminative variables}

\item{similarityThreshold}{to identify the simmilarity in cosine distance, this parameter
is the cut of simmilarity. Above this threshold we consider the variables highly similar.
Ex: 0.75, or 0.85}

\item{nameExperiment}{The name of the experiment together with the number of variables we included.
if is all model, allthe variables. Instead if is the selected model is less variables.
ex:
\enumerate{
\item nameExperiment <- paste0(name,"_", length(colnames(phenoSocial_4analysis)), " variables"); #all variables
\item nameExperiment <- paste0(name,"_", length(colnames(phenoAll_4analysis_woHighCor_sel1)), " variables"); · without highly correlated
\item nameExperiment <- paste0(name,"_", " more informative variables ",length(colnames(parameters_more_30perc_informative_Genotype)) ) #selected variables
}}
}
\value{
calculation of cosine simmilarity distances
}
\description{
Cosine Distance - Function to calculate cosine simmilarity distances
}
\examples{
similarityThreshold <- 0.86
stats_cosineDistance_plot1.function( catVarName,similarityThreshold, nameExperiment);

}
